feat: handling liff id for mini
