+++ .env
@@ -1,1 +1,4 @@
 VITE_LIFF_ID=1656508316-k7jNojXm
+
+# TODO
+VITE_LIFF_ID_MINI=
