+++ src/FilterTypes.ts
@@ -1,0 +1,4 @@
+export const FilterTypes = {
+  LIFF: 'LIFF',
+  MINI: 'MINI'
+} as const
