+++ src/vite-env.d.ts
@@ -2,6 +2,7 @@
 
 interface ImportMetaEnv {
   readonly VITE_LIFF_ID: string;
+  readonly VITE_LIFF_ID_MINI: string;
 }
 
 interface ImportMeta {
